+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+
+### Fixed
+
+
+## [1.3.0] - 2023-01-18
+
+### Added
+
+### Changed
 - template: conftest.py: Use session scope for firmware and backend fixtures
 - template: conftest.py: Targeted device must be specified with new `--device` required option.
 - unified naming of the Ledger Stax product: from `fat` or `fatstacks` to stax